Project 2/Student Management System RAG/requirements.txt
# Core Web Framework
streamlit==1.28.1

# LLM & RAG (all compatible versions)
langchain==0.1.20
langchain-openai==0.1.1
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-core>=0.1.7,<0.2

# Embeddings
sentence-transformers==2.2.2

# Vector Store
chromadb==0.5.0

# PDF Processing
pypdf==3.17.1

# Data Handling
python-dotenv==1.0.0
openai>=1.10.0,<2.0.0

# Utilities
python-dateutil==2.8.2
numpy==1.25.2
pydantic==2.5.0

# Additional dependencies
requests==2.31.0
urllib3==2.1.0

mysql-connector-python
